Update RR_config.js
Minor change such that this is able to scout more than the florida regionals with blue alliance integration. Look for the event name on blue alliance, the event input is the end part of the url to each event.  Such as the 2022hiho part of: https://www.thebluealliance.com/event/2022hiho. The previous default code was: 

      "Event":{
        "code":"e",
        "type":"event",
        "defaultValue":"2022flwp",
        "required":"true",
        "disabled":"true"